refactor(app-state): #245 facade/ 重命名为 ops/
facade(外观/桥接层)的命名暗示这些方法是「转发到别处」的中间层,与
「app-state 作为状态存储层、Core 操作经 CoreManager」的最终架构矛盾。

实际上这些是 TiangongState 自身的方法实现(会话 CRUD、持久化、配置缓存、
草稿管理等 app-state 核心职责),不是桥接层。重命名为 ops/(operations)
去掉错误语义,保留按职责分文件的结构。

同时 agent_config_facade.rs → agent_config.rs(去掉 _facade 后缀)。

纯重命名,零行为变更。
